+++ src/components/chat/ChatInterface.vue
@@ -24,7 +24,6 @@
         "
         :initial-sessions-open="!isFreshUser"
         :initial-explore-agents-open="!isFreshUser"
-        :initial-collections-open="!isFreshUser"
         :selected-session="sessionId"
         :add-dummy-session="Object.keys(conversations).length === 0"
         :selected-collection="collectionId"
@@ -32,7 +31,9 @@
         :sessions="sessions"
         :collections="collections"
         @create-new-session="createNewSession"
+        @create-collection="showCreateCollectionModal = true"
         @delete-session="showDeleteSessionDialog"
+        @delete-collection="promptDeleteCollection"
         @agent-click="
           if (!chatLoading) {
             handleTagAgent($event, false);
@@ -164,31 +165,47 @@
       </div>
     </div>
 
+    <NotificationCenter ref="notificationCenterRef" />
+
     <!-- Delete Session Dialog -->
-    <DeleteSessionDialog
+    <ConfirmModal
+      heading-text="Delete Session"
+      description-text="Are you sure you want to delete this session?"
+      action-button-text="Delete"
+      cancel-button-text="Cancel"
       :show-dialog="showDeleteDialog"
       @cancel-delete="
         showDeleteDialog = false;
         sessionToDelete = null;
       "
       @confirm-delete="confirmDeleteSession"
-    ></DeleteSessionDialog>
-
-    <!-- Success Banner -->
-    <SuccessBanner
-      v-if="showSuccessBanner"
-      :message="bannerMessage"
-      @hide="showSuccessBanner = false"
     />
 
     <!-- Delete Video Dialog -->
-    <DeleteVideoDialog
+    <ConfirmModal
+      heading-text="Delete Video"
+      description-text="Are you sure you want to delete this video? This action cannot be undone."
+      action-button-text="Delete"
+      cancel-button-text="Cancel"
       :show-dialog="showDeleteVideoDialog"
       @cancel-delete="
         showDeleteVideoDialog = false;
         videoToDelete = null;
       "
       @confirm-delete="confirmDeleteVideo"
+    />
+
+    <!-- Create Collection Modal-->
+    <CreateCollectionModal
+      :showDialog="showCreateCollectionModal"
+      @cancel="showCreateCollectionModal = false"
+      @create="promptCreateCollection"
+    />
+
+    <!-- Delete Collection Error Modal -->
+    <DeleteCollectionErrorModal
+      :isVisible="showDeleteErrorModal"
+      @closeModal="showDeleteErrorModal = false"
     />
 
     <!-- Upload Dialog -->
@@ -216,10 +233,14 @@ import SetupScreen from "./elements/SetupScreen.vue";
 import Sidebar from "./elements/Sidebar.vue";
 import UploadNotifications from "./elements/UploadNotifications.vue";
 import UploadVideoQueryCard from "./elements/UploadVideoQueryCard.vue";
+import NotificationCenter from "./elements/NotificationCenter.vue";
 
 import CollectionHeader from "./elements/CollectionHeader.vue";
+import ConfirmModal from "../modals/ConfirmModal.vue";
 import UploadModal from "../modals/UploadModal.vue";
-import DeleteSessionDialog from "../modals/DeleteSessionModal.vue";
+import DeleteCollectionErrorModal from "../modals/DeleteCollectionErrorModal.vue";
+import CreateCollectionModal from "../modals/CreateCollectionModal.vue";
+
 import ChatSearchResults from "../message-handlers/ChatSearchResults.vue";
 import ChatVideo from "../message-handlers/ChatVideo.vue";
 import ChatVideos from "../message-handlers/ChatVideos.vue";
@@ -229,11 +250,10 @@ import TextResponse from "../message-handlers/TextResponse.vue";
 import CollectionIcon from "../icons/Collection.vue";
 import DirectorIcon from "../icons/Director.vue";
 import ExternalLink from "../icons/ExternalLink.vue";
-import FileUploadIcon from "../icons/FileUpload.vue";
 import SearchIcon from "../icons/SearchIcon.vue";
 import QueryIcon from "../icons/Query.vue";
-import DeleteVideoDialog from "../modals/DeleteVideoModal.vue";
-import SuccessBanner from "../atoms/SuccessBanner.vue";
+import DeleteIcon from "../icons/Delete3.vue";
+import CheckIcon from "../icons/Check.vue";
 
 const props = defineProps({
   chatInputPlaceholder: {
@@ -318,6 +338,7 @@ const emit = defineEmits([]);
 const sidebarRef = ref(null);
 const chatInputRef = ref(null);
 const uploadNotificationsRef = ref(null);
+const notificationCenterRef = ref(null);
 
 const showCollectionView = ref(false);
 const taggedAgent = ref([]);
@@ -339,6 +360,7 @@ const {
   conversations,
   loadSession,
   uploadMedia,
+  createCollection,
   deleteCollection,
   refetchCollectionVideos,
   deleteVideo,
@@ -358,8 +380,7 @@ const chatWindowRef = ref(null);
 const headerRef = ref(null);
 const showDeleteVideoDialog = ref(false);
 const videoToDelete = ref(null);
-const showSuccessBanner = ref(false);
-const bannerMessage = ref("");
+const showDeleteErrorModal = ref(false);
 
 const isSetupComplete = computed(() => {
   return (
@@ -583,12 +604,61 @@ const confirmDeleteVideo = async () => {
 
   try {
     await deleteVideo(collection_id, id);
-    bannerMessage.value = "Video deleted successfully.";
-    showSuccessBanner.value = true;
+    notificationCenterRef.value.addNotification("Video deleted successfully.", {
+      type: "error",
+      icon: DeleteIcon,
+    });
   } catch (error) {
     console.error(`Error deleting video: ${error.message}`);
-    bannerMessage.value = "Error deleting video.";
-    showSuccessBanner.value = true;
+    notificationCenterRef.value.addNotification("Error deleting video", {
+      type: "error",
+      icon: DeleteIcon,
+    });
+  }
+};
+
+const showCreateCollectionModal = ref(false);
+
+const promptCreateCollection = async (newCollection) => {
+  showCreateCollectionModal.value = false;
+  try {
+    const createdCollection = await createCollection(
+      newCollection.name,
+      newCollection.description || " ",
+    );
+    notificationCenterRef.value.addNotification(
+      "Collection has been created successfully!",
+      {
+        type: "success",
+        icon: CheckIcon,
+      },
+    );
+  } catch (error) {
+    console.error("Error creating collection:", error.message);
+    notificationCenterRef.value.addNotification("Failed to create collection", {
+      type: "error",
+    });
+  }
+};
+
+const promptDeleteCollection = async (collection) => {
+  try {
+    await deleteCollection(collection?.id);
+    notificationCenterRef.value.addNotification(
+      "Collection deleted successfully.",
+      {
+        type: "error",
+        icon: DeleteIcon,
+      },
+    );
+  } catch (error) {
+    if (
+      error.message.includes("Invalid request: Your collection has non-zero")
+    ) {
+      showDeleteErrorModal.value = true;
+      return;
+    }
+    console.error("Unexpected error deleting collection:", error);
   }
 };
 
@@ -631,7 +701,6 @@ provide("videodb-chat", {
   setChatInput,
   registerMessageHandler,
   uploadMedia,
-  deleteCollection,
 });
 </script>